refactor(portfolio): move test files to tests/ directory
Separate test files from implementation by relocating all 16 server test files to a top-level tests/ directory. Convert relative imports to # alias, update vitest.config.ts (include/setupFiles/coverage.exclude), and move test fixtures alongside tests.
