+++ projects/portfolio/tests/features/chat-stub/chat-stub.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, vi } from 'vitest'
-import { chatStub } from './chat-stub'
+import { chatStub } from '#/server/features/chat-stub/chat-stub'
 
 describe('chatStub.completions', () => {
   it('completion の形で content を返す', async () => {
